Ajustes na tela de profile

+++ lib/model/carteira.dart
@@ -1,5 +1,5 @@
 class Carteira {
-  int saldo;
+  double saldo;
   String id;
 
   Carteira({
